requirements.txt
Click==7.0
Flask==1.0.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mongoengine==0.17.0
pymongo==3.7.2
python-dotenv==0.10.1
python-slugify==3.0.2
six==1.12.0
text-unidecode==1.2
Werkzeug==0.15.2
WTForms==2.2.1
